+++ jackson/src/main/scala/com/twitter/finatra/json/JsonLogging.scala
@@ -17,7 +17,7 @@ trait JsonLogging extends Logging {
   protected var _mapper: ScalaObjectMapper = _
 
   /* If JsonLogging is used w/ an Object, injection will not occur and we use the default mapper */
-  private lazy val logMapper = Option(_mapper) getOrElse JsonLogging.DefaultMapper
+  private lazy val logMapper = Option(_mapper).getOrElse(JsonLogging.DefaultMapper)
 
   /* Protected */